Key recipe holders and take integer directions
Direction.getNearest is integer-only now, so the vector components cast.
RecipeHolder is keyed by a ResourceKey rather than an identifier.

Error sites 1530 -> 1510.

+++ src/main/java/com/simibubi/create/content/fluids/drain/ItemDrainBlock.java
@@ -87,7 +87,7 @@ public class ItemDrainBlock extends Block implements IWrenchable, IBE<ItemDrainB
 		Vec3 deltaMovement = entityIn.getDeltaMovement()
 			.multiply(1, 0, 1)
 			.normalize();
-		Direction nearest = Direction.getNearest(deltaMovement.x, deltaMovement.y, deltaMovement.z, Direction.UP);
+		Direction nearest = Direction.getNearest((int) deltaMovement.x, (int) deltaMovement.y, (int) deltaMovement.z, Direction.UP);
 		ItemStack remainder = inputBehaviour.handleInsertion(itemEntity.getItem(), nearest, false);
 		itemEntity.setItem(remainder);
 		if (remainder.isEmpty())
